bs2ramfs: Use new Inode wrapper
This also replaces some uses of bindings::mode_t with the Mode type from
the kernel crate.

There are still some things that require another indirection to be
ported to Rust types, but most of the code we have now uses the new
type. I think that wrapping the `i_mapping` field could be difficult to
get right.