kernel crate: add ability to use predefined vtables from kernel
This moves the shared behavior of the `build` method from the file
operations vtable into its own trait, so that other types can also
implement it. Namely, we added the unit struct `SimpleDirOperations`
that returns a ref to the C global.

This could be expanded, I can see a general `trait BuildVtable<Table>`
and a macro like

    declare_c_vtable!(
        SimpleDirOperations,
        bindings::file_operations,
        bindings::simple_dir_operations,
    );

that expands to

    struct SimpleDirOperations;
    impl BuildVtable<bindings::file_operations> for SimpleDirOperations {
        fn build() -> &'static bindings::file_operations {
            &bindings::simple_dir_operations
        }
    }

so that we can easily make these constants accessible. The downside is,
that the interface for the caller is more abstract / complicated, since
the vtable struct has to be constructed on the calling site

    .set_file_operations::<FileOperationsVtable<NopFileOpenAdapter, Bs2RamfsFileOps>>();

We could however aid with some helper methods

    .set_file_ops_from_trait::<Bs2RamfsFileOps>();
    // ...
    fn set_file_ops_from_trait<Ops: FileOperations<A::Arg>, A: FileOpenAdapter = NopFileOpenAdapter>() { ... }

Implementing the `BuildVtable` trait for all implementors of
`FileOperations` is not possible, because there might be multiple types
`A` that satisfy the requirements, thus making a definite choice
impossible.