Use impl_flag_methods for FileSystemFlags
We already have a macro for flag types in
ec615990a409358c6adde319f3790b53758daed4

We could also try to make this as one trait that gives you free impls of
BitAnd and BitOr etc., and maybe add a derive(BitFlag) or so (we would
need one method that returns self.0 on these types). We could then also
have an associated type with the trait, so that we can also get rid of
ModeInt. I don't know how much we loose on the "const" side of things
with a trait though.